Always end script commands with a newline
Otherwise script will not be interpreted by the robot which might be counter-intuitive.
Changing the behavior as such will also be the same as in the ur_modern_driver
so migrating will be easier.

I decided to change the function's interface to copy the string in order add a
trailing '\n' if necessary.