Close all closable sockets
Sockets do not necessarily have to be in state connected when they should be
closed. Before, only connected sockets got closed leading to a "socket leak"
if a socket was disconnected before a close request was processed.

With this fix all sockets with a valid file descriptor get closed when close()
is being called.