Template all comm objects with the actual package type, not the header type
When designing this driver we wanted to have all communication objects inherit
from one common `Package` class.

As we want to serve two different protocols (RTDE and Primary/Secondary), we
had this Package class templated with a header type which is different in the
two protocols. With this design decision we could have one common communication
structure (Streams, Pipelines, Producers, Parsers, Consumers) without rewriting
code.

As the thing distinguishing the different protocols was the Header, we
decided to template all the communication objects using HeaderT.

However, as I recently realized, this destroys the possibility to easily create
consumers using the visitor pattern as being done in the `ur_modern_driver`.
With this, there would have to be one root consumer providing abstract methods
for all packages available (over all interfaces).

By templating the communication layer with the type of the actual package
(In terms of RTDEPackage or PrimaryPackage) we can establish a visitor pattern
at protocol level.