Initialize member variables
It can happen that the action gets triggered before the mode callback got triggered

While this changes stops the helper from crashing when this happens, it might
not be the best idea to do so as the question remains, what we should do
if we haven't even received a current status from the robot.

With the changes introduced inside this commit, the helper would trigger the
respective state changes, which might lead to wrong requests if we aren't entirely
sure what to do.

One solution would be to reject goals as long as no status was received,
but that would break such scenarios where you want to activate the robot automatically
during startup.

Another idea would be to delay actually starting the action server until we
received both, robot mode and safety mode. But I am not entirely sure whether
this will scale well.