driver: use default rate for JTC goal monitor.
The old values overrode the default of 20 Hz, which is low and leads to a worst-case delay of approx 100 ms between a goal state change and action clients being notified of that change.

This restores the rate to the default of 20 Hz.

If a higher update-rate would be desirable for a particular application, users should change it in their own configuration of the controllers.