Disable trajectory test for now
The trajectory test seems to not work anymore since a
couple of weeks. Running those locally (also with a
ursim running inside a docker container) works perfectly
fine, but running it inside the github action not.

As this is blocking many merges currently, I suggest
to disable this temporarily while opening an issue to fix it.