Use catkin_install_python macro for python files
This macro works just like the normal `install` macro, but it also
automatically changes the shebang line in the python file to `python2`
or `python3`, depending which version is used.

See: http://wiki.ros.org/UsingPython3/SourceCodeChanges#Changing_shebangs

What this means is that this package can be used with Python3 without
any further changes, for example in ROS Noetic.